+++ packages/fluentui/accessibility/src/behaviors/index.ts
@@ -3,6 +3,7 @@ export { default as alertWarningBehavior } from './Alert/alertWarningBehavior';
 export { default as alertBaseBehavior } from './Alert/alertBaseBehavior';
 export { default as attachmentBehavior } from './Attachment/attachmentBehavior';
 export { default as buttonBehavior } from './Button/buttonBehavior';
+export * from './Button/buttonBehavior';
 export { default as toggleButtonBehavior } from './Button/toggleButtonBehavior';
 export * from './Image/imageBehavior';
 export { default as imageBehavior } from './Image/imageBehavior';